common: move object_store to module
